add horizontal swiping navigation and render sessions per track
