+++ package.json
@@ -44,6 +44,7 @@
     "react": "16.0.0-alpha.12",
     "react-native": "^0.48.4",
     "react-native-elements": "^0.17.0",
+    "react-native-swiper": "^1.5.13",
     "react-navigation": "^1.0.0-beta.13"
   },
   "repository": {
