+++ screens/home/HomeContainer.js
@@ -1,9 +1,17 @@
 import React from 'react';
+import Swiper from 'react-native-swiper';
 import HomeScreen from './HomeScreen';
 import sessions from '../../assets/sessions.json';
+import tracks from '../../assets/tracks.json';
 
 const HomeContainer = () => (
-  <HomeScreen sessions={sessions} />
+  <Swiper showsButtons={false} loop={false} showsPagination={false} removeClippedSubviews={false} >
+    {
+      tracks.map(t => (
+        <HomeScreen key={t.id} trackName={t.title} trackId={t.id} sessions={sessions} />
+      ))
+    }
+  </Swiper>
 );
 
 export default HomeContainer;
